concert-ticketing/grpc/user-service/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
pika==1.3.1
python-dotenv==1.0.0
requests==2.32.0
sqlalchemy
alembic
bcrypt
PyJWT
grpcio
grpcio-tools
SQLAlchemy